+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+* Fixed an issue where **format** did not update the test playbook from its pack.
 * Fixed an issue where **update-release-notes** did not identified old yml integrations and scripts.
 * Added revision templates to the **update-release-notes** command.
 * Fixed an issue where **update-release-notes** crashed when a file was renamed.